python/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
dnspython==2.4.2
filelock==3.13.1
fonttools==4.47.0
fsspec==2023.10.0
idna==3.6
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
pymongo==4.6.1
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
sympy==1.12
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
typing_extensions==4.8.0
urllib3==2.1.0
